Fixed sway azerty bug.
When sway is launched the workspace switching doesn't work.
However this fix is provided in the sway config.
However the config change now only supports azerty.
Thus the azerty changes are present in the config_azerty file.
All we now need to do is hotswap the azerty file with the regular file
during boot. (check if the keyboard layout is azerty) otherwise we
hotswap the qwerty config file