Added extended Dark/Light theming
You can now select and change the theming more easily.
By setting the background property to light in the colors.conf
configuration file you can enable light mode (all other options are dark
mode)

When light mode is enabled each icon/image will be replaced by a dark
variant. This happens as followed.

Say we have a icon called icon.svg normally this icon will be loaded in.
When light mode is enable we search for icon.dark.svg instead.
The reason we look for a dark image is because the background is light.
Thus the foreground must be dark.

If no icon.dark.svg is found then we default to icon.svg instead.
Note that this option most likely won't provide good contrast but it is
a fallback