Don't crash the application switcher when no client is in focus
Previously we switched workspaces on the screen of the active client.
However when no client was in focus we couldn't switch to a different
tag because no screen was found.

However now we look for the screen that our mouse is on.
This way we can still travel to different tags but don't depend on a
active screen anymore.